refactor: centralize loja and blog components
