+++ app/blog/post/[slug]/page.tsx
@@ -6,8 +6,8 @@ import { isExternalUrl } from '@/utils/isExternalUrl'
 import type { Metadata } from 'next'
 import { getRelatedPostsFromPB } from '@/lib/posts/getRelatedPostsFromPB'
 import { getPostBySlug } from '@/lib/posts/getPostBySlug'
-import NextPostButton from '@/app/blog/components/NextPostButton'
-import PostSuggestions from '@/app/blog/components/PostSuggestions'
+import NextPostButton from '@/components/molecules/NextPostButton'
+import PostSuggestions from '@/components/organisms/PostSuggestions'
 import Script from 'next/script'
 
 export const dynamic = 'force-dynamic'
