+++ components/molecules/index.ts
@@ -2,3 +2,4 @@ export { FormField } from './FormField'
 export { InputWithMask } from './InputWithMask'
 export { default as SmoothTabs } from './SmoothTabs'
 export { default as SaldoCard } from './SaldoCard'
+export { default as NextPostButton } from './NextPostButton'
