+++ components/organisms/BlogHeroCarousel.tsx
